[lad-new-functions-plugin] rename hooks file to commands file.
